Fix some build warnings (#5941)
* Fix some build warnings

I added the -Wno-c99-designator flag because we were getting a warning
about syntax that requires c99 compatibility (which I guess isn't
exactly a subset of C++20). However, as far as I can tell, clang and gcc
support the syntax, so the warning isn't very useful to us at this point
in time.

* [run CI]

+++ bodo/libs/_groupby_col_set.cpp
@@ -357,7 +357,7 @@ void VarStdColSet::alloc_update_columns(
 
     // Add every value to update cols, except the first one
     //(the output column)
-    for (int i = init_start; i < out_cols.size(); i++) {
+    for (size_t i = init_start; i < out_cols.size(); i++) {
         this->update_cols.push_back(out_cols.at(i));
     }
 }
@@ -384,7 +384,7 @@ void VarStdColSet::alloc_combine_columns(
 
     // Add every value to update cols, except the first one
     //(the output column)
-    for (int i = 1; i < out_cols.size(); i++) {
+    for (size_t i = 1; i < out_cols.size(); i++) {
         this->combine_cols.push_back(out_cols.at(i));
     }
 }
@@ -530,7 +530,7 @@ void SkewColSet::alloc_update_columns(
 
     // Add every value to update cols, except the first one
     //(the output column)
-    for (int i = init_start; i < out_cols.size(); i++) {
+    for (size_t i = init_start; i < out_cols.size(); i++) {
         this->update_cols.push_back(out_cols.at(i));
     }
 }
@@ -558,7 +558,7 @@ void SkewColSet::alloc_combine_columns(
 
     // Add every value to update cols, except the first one
     //(the output column)
-    for (int i = 1; i < out_cols.size(); i++) {
+    for (size_t i = 1; i < out_cols.size(); i++) {
         this->combine_cols.push_back(out_cols.at(i));
     }
 }
@@ -1036,7 +1036,7 @@ void KurtColSet::alloc_update_columns(
 
     // Add every value to update cols, except the first one
     //(the output column)
-    for (int i = init_start; i < out_cols.size(); i++) {
+    for (size_t i = init_start; i < out_cols.size(); i++) {
         this->update_cols.push_back(out_cols.at(i));
     }
 }
@@ -1124,7 +1124,7 @@ void KurtColSet::alloc_combine_columns(
 
     // Add every value to update cols, except the first one
     //(the output column)
-    for (int i = 1; i < out_cols.size(); i++) {
+    for (size_t i = 1; i < out_cols.size(); i++) {
         this->combine_cols.push_back(out_cols.at(i));
     }
 }
